+++ openai-client/src/commonMain/kotlin/com.aallam.openai.client/internal/api/ImagesApi.kt
@@ -62,6 +62,7 @@ internal class ImagesApi(private val requester: HttpRequester) : Images {
         edit.n?.let { n -> append(key = "n", value = n) }
         edit.size?.let { dim -> append(key = "size", value = dim.size) }
         edit.user?.let { user -> append(key = "user", value = user) }
+        edit.model?.let { model -> append(key = "model", value = model.id) }
     }
 
     override suspend fun imageURL(variation: ImageVariation): List<ImageURL> {
@@ -89,16 +90,16 @@ internal class ImagesApi(private val requester: HttpRequester) : Images {
         edit.n?.let { n -> append(key = "n", value = n) }
         edit.size?.let { dim -> append(key = "size", value = dim.size) }
         edit.user?.let { user -> append(key = "user", value = user) }
-
+        edit.model?.let { model -> append(key = "model", value = model.id) }
     }
 
     /** Convert [ImageCreation] instance to base64 JSON request */
     private fun ImageCreation.toJSONRequest() = ImageCreationRequest(
-        prompt = prompt, n = n, size = size, user = user, responseFormat = ImageResponseFormat.base64Json,
+        prompt = prompt, n = n, size = size, user = user, responseFormat = ImageResponseFormat.base64Json, model = model?.id,
     )
 
     /** Convert [ImageCreation] instance to URL request */
     private fun ImageCreation.toURLRequest() = ImageCreationRequest(
-        prompt = prompt, n = n, size = size, user = user, responseFormat = ImageResponseFormat.url,
+        prompt = prompt, n = n, size = size, user = user, responseFormat = ImageResponseFormat.url, model = model?.id
     )
 }