+++ openai-core/src/commonMain/kotlin/com.aallam.openai.api/image/ImageCreation.kt
@@ -2,6 +2,7 @@ package com.aallam.openai.api.image
 
 import com.aallam.openai.api.BetaOpenAI
 import com.aallam.openai.api.OpenAIDsl
+import com.aallam.openai.api.model.ModelId
 
 /**
  * Image generation request.
@@ -23,14 +24,20 @@ public class ImageCreation(
     /**
      * The format in which the generated images are returned. Must be one of url or b64_json.
      */
-    public val user: String? = null
+    public val user: String? = null,
+
+    /**
+     * The model used to generate image. Must be one of dall-e-2 or dall-e-3. If not provided, dall-e-2 is used.
+     */
+    public val model: ModelId? = null,
 )
 
 /**
  * Image generation request.
  */
 @BetaOpenAI
-public fun imageCreation(block: ImageCreationBuilder.() -> Unit): ImageCreation = ImageCreationBuilder().apply(block).build()
+public fun imageCreation(block: ImageCreationBuilder.() -> Unit): ImageCreation =
+    ImageCreationBuilder().apply(block).build()
 
 /**
  * Builder of [ImageCreation] instances.
@@ -60,12 +67,18 @@ public class ImageCreationBuilder {
     public var user: String? = null
 
     /**
+     * The model used to generate image. Must be one of dall-e-2 or dall-e-3. If not provided, dall-e-2 is used.
+     */
+    public var model: ModelId? = null
+
+    /**
      * Creates the [ImageCreation] instance
      */
     public fun build(): ImageCreation = ImageCreation(
         prompt = requireNotNull(prompt),
         n = n,
         size = size,
-        user = user
+        user = user,
+        model = model,
     )
 }