+++ openai-core/src/commonMain/kotlin/com.aallam.openai.api/image/ImageVariation.kt
@@ -3,6 +3,7 @@ package com.aallam.openai.api.image
 import com.aallam.openai.api.BetaOpenAI
 import com.aallam.openai.api.OpenAIDsl
 import com.aallam.openai.api.file.FileSource
+import com.aallam.openai.api.model.ModelId
 
 /**
  * Image variant request.
@@ -27,6 +28,11 @@ public class ImageVariation(
      * A unique identifier representing your end-user, which can help OpenAI to monitor and detect abuse.
      */
     public val user: String? = null,
+
+    /**
+     * The model used to generate image. Must be one of dall-e-2 or dall-e-3. If not provided, dall-e-2 is used.
+     */
+    public val model: ModelId? = null,
 )
 
 /**
@@ -63,12 +69,18 @@ public class ImageVariationBuilder {
     public var user: String? = null
 
     /**
+     * The model used to generate image. Must be one of dall-e-2 or dall-e-3. If not provided, dall-e-2 is used.
+     */
+    public var model: ModelId? = null
+
+    /**
      * Creates the [ImageVariation] instance
      */
     public fun build(): ImageVariation = ImageVariation(
         image = requireNotNull(image) { "image is required" },
         n = n,
         size = size,
-        user = user
+        user = user,
+        model = model,
     )
 }